requirements.txt
streamlit
sounddevice
numpy
scipy
openai-whisper
gtts
google-genai

